Create controller for places routes
