+++ models/http-error.model.js
@@ -1,4 +1,4 @@
-export class HttpError extends Error {
+export default class HttpError extends Error {
   constructor(message, errorCode) {
     super(message); // Add a "message" property
     this.code = errorCode; // Adds a "code" property
